move addDonut function to its own module
