+++ test/tests.js
@@ -1,13 +1,6 @@
+import addDonut from '../src/add-donut.js'
 const test = QUnit.test;
-
-function addDonut(parentElement, donutFlavor) {
-    const span = document.createElement('span');
-    span.classList.add('part');
-    span.classList.add(donutFlavor);
-
-    parentElement.appendChild(span);
-}
-
+ 
 test('append selected donut to dog element', function(assert) {
     //Arrange
     // Set up your parameters and expectations
@@ -22,4 +15,4 @@ test('append selected donut to dog element', function(assert) {
     //Assert
     const parentInnerHTML = dogParentElement.innerHTML; 
     assert.equal(parentInnerHTML, expected);
-});
+})
